requirements.txt
aiodns==3.0.0
argh==0.26.2
beautifulsoup4==4.9.3
Django==3.2.23
django-ace==1.0.11
django-celery-beat==2.2.1
django-login-required-middleware==0.6.1
django-mathfilters==1.0.0
django-role-permissions==3.2.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
dotted-dict==1.1.3
drf-yasg==1.21.3
gevent==23.9.1
gunicorn==21.2.0
humanize==4.3.0
Markdown==3.3.4
metafinder==1.2
netaddr==0.8.0
netlas==0.4.1
openai==0.28.0
psycopg2==2.9.7
pycvesearch==1.0
PySocks==1.7.1
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
scapy==2.4.3
tldextract==3.5.0
uro==1.0.0
validators==0.18.2
wafw00f==2.2.0
watchdog==3.0.0
weasyprint==53.3
whatportis
xmltodict==0.13.0
